Got Random Spawning of enemies,health, and astronauts done. UFO enemeies have been made. Menus have been added.

+++ Assets/Scripts/DamageOnOverlap.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class DamageOnOverlap : MonoBehaviour
 {
     public bool isInstaKill;
-    public float damageDone;
+    public int damageDone;
     public float scoreAmount;
